+++ CHANGELOG.md
@@ -15,7 +15,7 @@ ___Note:__ Yet to be released changes appear here._
 * `DEPS`: update to `@camunda/linting@3.48.1`
 * `DEPS`: update to `@bpmn-io/variable-resolver@1.4.1`
 * `DEPS`: update to `@bpmn-io/extract-process-variables@2.0.0`
-* `DEPS`: update to `camunda-bpmn-js@5.20.0`
+* `DEPS`: update to `camunda-bpmn-js@5.22.0`
 * `DEPS`: update to `bpmn-js@18.12.0`
 * `DEPS`: update to `bpmn-js-properties-panel@5.50.1`
 * `DEPS`: update to `bpmn-js-element-templates@2.20.0`
@@ -38,6 +38,7 @@ ___Note:__ Yet to be released changes appear here._
 
 ### BPMN
 
+* `FEAT`: add support of copy selection as image ([#5613](https://github.com/camunda/camunda-modeler/pull/5613), [#5592](https://github.com/camunda/camunda-modeler/issues/5592))
 * `FEAT`: add support of FEEL expressions for 'Form ID' field ([#5231](https://github.com/camunda/camunda-modeler/issues/5231))
 * `FEAT`: support conditional events modeling ([#5400](https://github.com/camunda/camunda-modeler/issues/5400))
 * `FEAT`: validate variable names ([#5188](https://github.com/camunda/camunda-modeler/issues/5188))